Refactor provider classes and update configuration paths
