+++ llmproxy/provider/google/vertexai.py
@@ -1,7 +1,7 @@
 from google.cloud import aiplatform
 from google.auth import exceptions as auth_exceptions
 from google.api_core import exceptions as api_exceptions
-from llmproxy.provider.base import BaseModel, CompletionResponse
+from llmproxy.provider.base import BaseProvider, CompletionResponse
 from llmproxy.utils.enums import BaseEnum
 from llmproxy.utils.log import logger
 from vertexai.language_models import TextGenerationModel
@@ -16,6 +16,23 @@ vertexai_price_data = {
     },
 }
 
+vertexai_category_data = {
+    "model-categories": {
+        "text-bison": {
+            "Code Generation Task": 1,
+            "Text Generation Task": 1,
+            "Translation and Multilingual Applications Task": 1,
+            "Natural Language Processing Task": 1,
+            "Conversational AI Task": 2,
+            "Educational Applications Task": 1,
+            "Healthcare and Medical Task": 2,
+            "Legal Task": 2,
+            "Financial Task": 2,
+            "Content Recommendation Task": 1,
+        },
+    }
+}
+
 
 class VertexAIModel(str, BaseEnum):
     # Add other models
@@ -23,7 +40,7 @@ class VertexAIModel(str, BaseEnum):
     PALM_CHAT = "chat-bison"
 
 
-class VertexAI(BaseModel):
+class VertexAI(BaseProvider):
     def __init__(
         self,
         prompt: str = "",
@@ -104,6 +121,13 @@ class VertexAI(BaseModel):
         logger.info(f"Calculated Cost: {cost}")
 
         return cost
+
+    def get_category_rank(self, category: str = "") -> str:
+        logger.info(msg=f"Current model: {self.model}")
+        logger.info(msg=f"Category of prompt: {category}")
+        category_rank = vertexai_category_data["model-categories"][self.model][category]
+        logger.info(msg=f"Rank of category: {category_rank}")
+        return category_rank
 
     def _handle_error(self, exception: str, error_type: str) -> CompletionResponse:
         return CompletionResponse(message=exception, err=error_type)
