+++ llmproxy/provider/openai/chatgpt.py
@@ -1,4 +1,4 @@
-from llmproxy.provider.base import BaseModel, CompletionResponse
+from llmproxy.provider.base import BaseProvider, CompletionResponse
 from llmproxy.utils.enums import BaseEnum
 from llmproxy.utils.log import logger
 import openai
@@ -39,6 +39,59 @@ open_ai_price_data = {
     },
 }
 
+open_ai_category_data = {
+    "model-categories": {
+        "gpt-3.5-turbo-1106": {
+            "Code Generation Task": 2,
+            "Text Generation Task": 1,
+            "Translation and Multilingual Applications Task": 2,
+            "Natural Language Processing Task": 1,
+            "Conversational AI Task": 1,
+            "Educational Applications Task": 2,
+            "Healthcare and Medical Task": 3,
+            "Legal Task": 3,
+            "Financial Task": 3,
+            "Content Recommendation Task": 2,
+        },
+        "gpt-3.5-turbo-instruct": {
+            "Code Generation Task": 2,
+            "Text Generation Task": 1,
+            "Translation and Multilingual Applications Task": 2,
+            "Natural Language Processing Task": 1,
+            "Conversational AI Task": 1,
+            "Educational Applications Task": 2,
+            "Healthcare and Medical Task": 3,
+            "Legal Task": 3,
+            "Financial Task": 3,
+            "Content Recommendation Task": 2,
+        },
+        "gpt-4": {
+            "Code Generation Task": 1,
+            "Text Generation Task": 1,
+            "Translation and Multilingual Applications Task": 1,
+            "Natural Language Processing Task": 1,
+            "Conversational AI Task": 1,
+            "Educational Applications Task": 1,
+            "Healthcare and Medical Task": 2,
+            "Legal Task": 2,
+            "Financial Task": 2,
+            "Content Recommendation Task": 1,
+        },
+        "gpt-4-32k": {
+            "Code Generation Task": 1,
+            "Text Generation Task": 1,
+            "Translation and Multilingual Applications Task": 1,
+            "Natural Language Processing Task": 1,
+            "Conversational AI Task": 1,
+            "Educational Applications Task": 1,
+            "Healthcare and Medical Task": 2,
+            "Legal Task": 2,
+            "Financial Task": 2,
+            "Content Recommendation Task": 1,
+        },
+    }
+}
+
 
 class OpenAIModel(str, BaseEnum):
     GPT_4_1106_PREVIEW = "gpt-4-1106-preview"
@@ -49,7 +102,7 @@ class OpenAIModel(str, BaseEnum):
     GPT_3_5_TURBO_INSTRUCT = "gpt-3.5-turbo-instruct"
 
 
-class OpenAI(BaseModel):
+class OpenAI(BaseProvider):
     def __init__(
         self,
         prompt: str = "",
@@ -128,6 +181,13 @@ class OpenAI(BaseModel):
         logger.info(f"Calculated Cost: {cost}")
 
         return cost
+
+    def get_category_rank(self, category: str = "") -> str:
+        logger.info(msg=f"Current model: {self.model}")
+        logger.info(msg=f"Category of prompt: {category}")
+        category_rank = open_ai_category_data["model-categories"][self.model][category]
+        logger.info(msg=f"Rank of category: {category_rank}")
+        return category_rank
 
     def _handle_error(self, exception: str, error_type: str) -> CompletionResponse:
         return CompletionResponse(message=exception, err=error_type)
